Factor out need for ERB in JS files
The Babel transformation was booting up Rails just to get the env
which is slow and makes using Jest needlessly complicated.

These process references are dynamically replaced. See: https://cli.vuejs.org/guide/mode-and-env.html#using-env-variables-in-client-side-code
