+++ app/webpacker/store/index.js
@@ -8,7 +8,7 @@ import createLogger from "vuex/dist/logger";
 
 Vue.use(Vuex);
 
-const debug = <%= Rails.env.development? %>;
+const debug = process.env.NODE_ENV == "development";
 
 export default new Vuex.Store({
   modules: {annotations,
